+++ src/apis/user/__init__.py
@@ -38,6 +38,8 @@ routes = [
     APIRoute("/user/{uid}/accept", manage.post_accept, methods=["POST"], response_class=JSONResponse),
     APIRoute("/user/{uid}/discord", manage.patch_discord, methods=["PATCH"], response_class=JSONResponse),
     APIRoute("/user/{uid}/connections", manage.delete_connections, methods=["DELETE"], response_class=JSONResponse),
+    APIRoute("/user/ban/list", manage.get_ban_list, methods=["GET"], response_class=JSONResponse),
+    APIRoute("/user/ban", manage.get_ban, methods=["GET"], response_class=JSONResponse),
     APIRoute("/user/ban", manage.put_ban, methods=["PUT"], response_class=JSONResponse),
     APIRoute("/user/ban", manage.delete_ban, methods=["DELETE"], response_class=JSONResponse),